refactor: Réorganise les composants de recherche APE
