+++ site/source/components/GuichetInfo.tsx
