+++ site/source/pages/assistants/recherche-code-ape/components/HelpGuichetUnique.tsx
@@ -1,0 +1,26 @@
+import { HelpButtonWithPopover } from '@/design-system/buttons'
+import { Link } from '@/design-system/typography/link'
+import { Body } from '@/design-system/typography/paragraphs'
+
+export function HelpGuichetUnique() {
+	return (
+		<HelpButtonWithPopover
+			type="info"
+			title="Qu'est-ce que le guichet unique ?"
+		>
+			<Body>
+				Le{' '}
+				<Link href="https://procedures.inpi.fr/">
+					Guichet électronique des formalités d’entreprises
+				</Link>{' '}
+				(Guichet unique) est un portail internet sécurisé, auprès duquel toute
+				entreprise est tenue de déclarer sa création, depuis le 1er janvier
+				2023.
+			</Body>
+			<Body>
+				Il utilise une classification des activités différente de celle utilisée
+				par l'INSEE pour code APE.
+			</Body>
+		</HelpButtonWithPopover>
+	)
+}
